[ignore] Optimize for loops for ndo_port_channel_interface and Move lookup_valid_interfaces function to mso.py from module_utils.
